+++ commitlint.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   extends: ['@commitlint/config-conventional'],
   rules: {
-    'type-enum': () => [2, 'always', ['fix', 'feat', 'chore', 'refactor', 'docs', 'test']],
+    'type-enum': () => [2, 'always', ['fix', 'feat', 'chore', 'refactor', 'docs', 'test', 'ci']],
     'body-max-line-length': () => [2, 'always', 500],
   },
 };